requirements.txt
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.3
greenlet==3.0.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipython==8.22.2
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jmespath==1.0.1
lxml==5.1.0
matplotlib-inline==0.1.6
packaging==24.0
parsel==1.9.0
parso==0.8.3
prompt-toolkit==3.0.43
Protego==0.3.0
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.17.2
pyOpenSSL==24.1.0
python-dotenv==1.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
stack-data==0.6.3
tldextract==5.1.2
traitlets==5.14.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.10.0
urllib3==2.2.1
w3lib==2.1.2
wcwidth==0.2.13
zope.interface==6.2
